feat(pipe): get userDto using Bearer token

+++ 이기웅/prisma-post-user/src/modules/post/post.controller.ts
@@ -6,26 +6,39 @@ import {
 import {
   ResponseInterceptor,
   ResponseWithIdInterceptor,
+  Roles,
 } from '@/common/interceptors';
 import {
   Body,
   Controller,
   Delete,
+  ForbiddenException,
   Get,
   HttpCode,
   Param,
   Patch,
   Post,
   Query,
+  UseGuards,
   UseInterceptors,
 } from '@nestjs/common';
-import { ApiOperation, ApiResponse, ApiTags } from '@nestjs/swagger';
+import {
+  ApiBearerAuth,
+  ApiOperation,
+  ApiResponse,
+  ApiTags,
+} from '@nestjs/swagger';
 import { UserService } from '../user/user.service';
 import { CreatePostDto } from './dto/create-post.dto';
 import { GetPostsQueryDto } from './dto/get-posts-query.dto';
 import { PostDto } from './dto/post.dto';
 import { UpdatePostDto } from './dto/update-post.dto';
 import { PostService } from './post.service';
+import { ACCESS_TOKEN } from 'src/utils/constants/jwt.constant';
+import { JwtAuthGuard } from '@/common/guards';
+import { ROLE } from 'src/utils/constants';
+import { User } from '@/common/decorators';
+import { UserDto } from '../user/dto/user.dto';
 
 @ApiTags('Post')
 @Controller('posts')
@@ -70,18 +83,32 @@ export class PostController {
   @ApiOperation({ summary: '게시물 생성' })
   @UseInterceptors(ResponseWithIdInterceptor)
   @HttpCode(201)
-  async createPost(@Body() createPostDto: CreatePostDto) {
-    return await this.postService.createPost(createPostDto);
+  @ApiBearerAuth(ACCESS_TOKEN)
+  @UseGuards(JwtAuthGuard)
+  @Roles(ROLE.ADMIN, ROLE.USER)
+  async createPost(
+    @Body() createPostDto: CreatePostDto,
+    @User() user: UserDto,
+  ) {
+    return await this.postService.createPost(createPostDto, user.id);
   }
 
   @Patch(':id')
   @ApiResponse({ status: 204 })
   @ApiOperation({ summary: 'postId로 게시물 수정' })
   @HttpCode(204)
+  @ApiBearerAuth(ACCESS_TOKEN)
+  @UseGuards(JwtAuthGuard)
+  @Roles(ROLE.ADMIN, ROLE.USER)
   async updatePost(
     @Param('id') id: string,
     @Body() updatePostDto: UpdatePostDto,
+    @User() user: UserDto,
   ) {
+    if (user.role !== ROLE.ADMIN && id !== user.id) {
+      throw new ForbiddenException('권한이 없습니다.');
+    }
+
     return await this.postService.updatePost(id, updatePostDto);
   }
 
@@ -89,7 +116,14 @@ export class PostController {
   @ApiResponse({ status: 204 })
   @ApiOperation({ summary: 'postId로 게시물 삭제' })
   @HttpCode(204)
-  async deletePost(@Param('id') id: string) {
+  @ApiBearerAuth(ACCESS_TOKEN)
+  @UseGuards(JwtAuthGuard)
+  @Roles(ROLE.ADMIN, ROLE.USER)
+  async deletePost(@Param('id') id: string, @User() user: UserDto) {
+    if (user.role !== ROLE.ADMIN && id !== user.id) {
+      throw new ForbiddenException('권한이 없습니다.');
+    }
+
     return await this.postService.deletePost(id);
   }
 }
